DEV-2687: Keep isTouchSynthesizedMouseEvent out of the public Handsontable.dom namespace
