+++ handsontable/src/3rdparty/walkontable/AGENTS.md
@@ -186,7 +186,7 @@ Never guess a container from an empty style read, and never cache a layout decis
 
 `isMobileBrowser()` reads the user agent, `isTouchSupported()` reads `'ontouchstart' in window`. iPad Safari/Chrome (desktop UA since iPadOS 13) and Windows touchscreens are **desktop UA + touch**, so `event.ts` registers both listener sets and the browser's synthesized `mousedown`/`mouseup`/`click` sequence after every `touchend` reaches the mouse listeners. Rules:
 
-- **Drop the synthesized sequence symmetrically.** `#isTouchSynthesizedMouseEvent()` gates both `mousedown` and `mouseup`. Dropping only `mouseup` (the original #12804 fix) fired `onCellMouseDown` twice per tap and re-armed the pairing slot and made iPad double-tap-to-edit nondeterministic (DEV-2687). Detection: Blink `sourceCapabilities.firesTouchEvents` via `helpers/dom/event.ts` `isTouchSynthesizedMouseEvent()`; WebKit/Firefox fall back to the 500 ms window after the last touch-driven `onMouseUp` (`#lastTouchMouseUpAt`). `tableView.ts` uses the same helper for the outside-click handler.
+- **Drop the synthesized sequence symmetrically.** `#isTouchSynthesizedMouseEvent()` gates both `mousedown` and `mouseup`. Dropping only `mouseup` (the original #12804 fix) fired `onCellMouseDown` twice per tap and re-armed the pairing slot and made iPad double-tap-to-edit nondeterministic (DEV-2687). Detection: Blink `sourceCapabilities.firesTouchEvents` via `helpers/dom/inputOrigin.ts` `isTouchSynthesizedMouseEvent()`; WebKit/Firefox fall back to the 500 ms window after the last touch-driven `onMouseUp` (`#lastTouchMouseUpAt`). `tableView.ts` uses the same helper for the outside-click handler.
 - **Touch double-tap window is 1000 ms** (`TOUCH_DBLTAP_TIMEOUT`, equal to the mousedown slot). A tap arms both `#dblClickOrigin` slots at `touchend`, so the 500 ms mouseup slot was the binding limit; iOS drops taps under ~250 ms and humans double-tap at 300–600 ms. Mouse double-click timing (1000/500) is unchanged. `TOUCH_SYNTHESIZED_MOUSE_WINDOW` (500 ms) is deliberately shorter than the pairing window, so on the WebKit fallback path a mouse pair arriving 500–1000 ms after a tap completes the double-click; see the constant's JSDoc for the invariants.
 - **Before 18.1, 18.0.0 opened the editor on every single tap** on these devices — the unsuppressed synthesized pair completed a second dblclick match. That was a side effect, not a feature; do not "restore" it.
 - Test it in Playwright with `test.use({ ...devices['Desktop Chrome'], hasTouch: true, browserName: 'chromium' })` (`tests/e2e/touch-tap-to-edit.spec.ts`); Chromium synthesizes the same mouse sequence after `locator.tap()`. Use `page.clock` to drive the pairing timers. Script-dispatched `MouseEvent`s have `sourceCapabilities === null` and exercise the WebKit fallback.
